Introduce IdSlice that is to Id what str is to String
This allows enforcing more typing around knowing that &str s are an id
or no.

Given that we currently have CoercesTo<Id> for str, which means we can
use &str directly in inputs though, this is probably not worth the added
complexity, so I'm going to leave it as "prototype" state.